Use rust toolchain channel 1.74.0
... to get access to new features in general and because using trunk is
easier on >=1.70

+++ crates/aocoracle/src/y2021/d07.rs
@@ -66,7 +66,7 @@ pub fn part_2(input: &str) -> Result<String, AnyError> {
     let cost = (min..=max)
         .map(|location| _quadratic_cost(&census, location))
         .min()
-        .unwrap() as u64;
+        .unwrap();
     Ok(format!("{}", cost))
 }
